move vote_collector to engines
